requirements.txt
sentence-transformers
langchain
pypdf
flask
python-dotenv
langchain-pinecone
langchain-community
langchain-experimental
langchain-openai
langchain-groq
pinecone-client
torchvision 
torchaudio
pinecone




groq
pinecone
tensorflow
transformers
tf-keras
torch